Remove TODO: split empowerment.py into empowerment.py, education.py and demographics.py
